FileAPI/tests/submissions/Opera/BlobBuilder.html

<!doctype html>
<html>
 <head>
  <title>BlobBuilder absence</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script>
  test(function(){
    var prefixes = ['', 'O', 'Moz', 'WebKit', 'MS'];
    for (var i = 0; i < prefixes.length; ++i) {
      assert_false(prefixes[i]+'BlobBuilder' in window, prefixes[i]+'BlobBuilder');
    }
  });
  </script>
 </body>
</html>
